feat: comprehensive architecture, performance, and safety overhaul - File format v2 with extended header (128 bytes): directory table for   incremental updates, extension bitmap index, and bi-gram index - External merge sort: chunks of 100K entries sorted to disk,   K-way merged via binary heap for constant memory indexing - True incremental updates via directory mtime comparison against   old index; falls back to full rebuild on v1 indexes - Bi-gram fallback for 2-char patterns (no more full scan) - Extension bitmap index for common file types - Regex literal extraction for trigram pre-filtering - Unicode-aware case folding with NFC normalization and caseless - Lock file prevents concurrent index writes - SIGINT handler cleans temp files on interrupt - SIGBUS handler provides clean error during race-condition reads - --version on mupdatedb, --verbose on both binaries - --schema kept as alias, primary long flag is --statistics - shlex-based PAGER argument parsing - Consistent type safety: mode is u32, size is u64 across all structs - Standardized error prefixes (mlocate: / mupdatedb:) - serde_json for count output formatting - New integration tests: --size filter, --type filter - New unit tests: Unicode folding, bigram generation, large chunk merge - 59 tests passing, zero clippy warnings

+++ src/pipeline.rs
@@ -6,7 +6,7 @@ pub struct FileEntry {
     pub full_path: String,
     pub size: u64,
     pub mtime: i64,
-    pub mode: i32,
+    pub mode: u32,
     pub mime_type: String,
 }
 
@@ -45,7 +45,7 @@ fn extract_metadata(path: &std::path::Path) -> Option<FileEntry> {
     use std::os::unix::fs::MetadataExt;
     let size = meta.len();
     let mtime = meta.mtime();
-    let mode = meta.mode() as i32;
+    let mode = meta.mode();
 
     let mime_type = detect_mime(path);
